Move size_hint module out of private::de
