+++ serde/src/private/mod.rs
@@ -1,5 +1,6 @@
 pub mod de;
 pub mod ser;
+pub mod size_hint;
 
 // FIXME: #[cfg(doctest)] once https://github.com/rust-lang/rust/issues/67295 is fixed.
 pub mod doc;
